+++ fbw-a32nx/src/systems/instruments/src/MCDU/legacy_pages/atsu/A320_Neo_CDU_ATC_ClearanceReq.ts
@@ -4,7 +4,7 @@ import { Keypad } from '../../legacy/A320_Neo_CDU_Keypad';
 import { CDUAtcDepartReq } from './A320_Neo_CDU_ATC_DepartReq';
 import { CDUAtcMenu } from './A320_Neo_CDU_ATC_Menu';
 import { CDUAtcTextFansA } from './FansA/A320_Neo_CDU_ATC_Text';
-import { LegacyAtsuPageInterface } from '../../legacy/LegacyAtsuPageInterface';
+import { LegacyAtsuPageInterface, setKeyNotActiveLskActions } from '../../legacy/LegacyAtsuPageInterface';
 
 export class CDUAtcClearanceReq {
   static CreateDataBlock() {
@@ -111,5 +111,6 @@ export class CDUAtcClearanceReq {
         CDUAtcClearanceReq.ShowPage(mcdu, title);
       }
     };
+    setKeyNotActiveLskActions(mcdu);
   }
 }