+++ fbw-a32nx/src/systems/instruments/src/MCDU/legacy_pages/atsu/FansA/A320_Neo_CDU_ATC_ContactRequest.ts
@@ -4,7 +4,7 @@ import { Keypad } from '../../../legacy/A320_Neo_CDU_Keypad';
 import { CDUAtcFlightReq } from '../A320_Neo_CDU_ATC_FlightReq';
 import { CDUAtcTextFansA } from '../FansA/A320_Neo_CDU_ATC_Text';
 import { NXSystemMessages } from '../../../messages/NXSystemMessages';
-import { LegacyAtsuPageInterface } from '../../../legacy/LegacyAtsuPageInterface';
+import { LegacyAtsuPageInterface, setKeyNotActiveLskActions } from '../../../legacy/LegacyAtsuPageInterface';
 
 export class CDUAtcContactRequest {
   static CreateDataBlock(): any {
@@ -41,6 +41,7 @@ export class CDUAtcContactRequest {
 
   static ShowPage(mcdu: LegacyAtsuPageInterface, data = CDUAtcContactRequest.CreateDataBlock()) {
     mcdu.clearDisplay();
+    mcdu.page.Current = mcdu.page.ATCContactRequest;
 
     let requestContact = '{cyan}{{end}REQ VOICE CONTACT';
     if (data.altitude) {
@@ -148,5 +149,6 @@ export class CDUAtcContactRequest {
         }
       }
     };
+    setKeyNotActiveLskActions(mcdu);
   }
 }