+++ fbw-a32nx/src/systems/instruments/src/MCDU/legacy_pages/atsu/FansA/A320_Neo_CDU_ATC_LatRequest.ts
@@ -11,7 +11,7 @@ import { Keypad } from '../../../legacy/A320_Neo_CDU_Keypad';
 import { CDUAtcFlightReq } from '../A320_Neo_CDU_ATC_FlightReq';
 import { CDUAtcTextFansA } from '../FansA/A320_Neo_CDU_ATC_Text';
 import { NXSystemMessages } from '../../../messages/NXSystemMessages';
-import { LegacyAtsuPageInterface } from '../../../legacy/LegacyAtsuPageInterface';
+import { LegacyAtsuPageInterface, setKeyNotActiveLskActions } from '../../../legacy/LegacyAtsuPageInterface';
 import { WaypointEntryUtils } from '@fmgc/flightplanning/WaypointEntryUtils';
 
 export class CDUAtcLatRequestFansA {
@@ -91,6 +91,7 @@ export class CDUAtcLatRequestFansA {
 
   static ShowPage1(mcdu: LegacyAtsuPageInterface, data = CDUAtcLatRequestFansA.CreateDataBlock()) {
     mcdu.clearDisplay();
+    mcdu.page.Current = mcdu.page.ATCLatRequest;
 
     let weatherDeviation = '{cyan}[  ]{end}';
     if (data.weatherDeviation) {
@@ -305,6 +306,7 @@ export class CDUAtcLatRequestFansA {
 
   static ShowPage2(mcdu: LegacyAtsuPageInterface, data = CDUAtcLatRequestFansA.CreateDataBlock()) {
     mcdu.clearDisplay();
+    mcdu.page.Current = mcdu.page.ATCLatRequest;
 
     let offsetDistance = '[  ]';
     let offsetStartPoint = '[     ]';
@@ -478,5 +480,6 @@ export class CDUAtcLatRequestFansA {
     mcdu.onNextPage = () => {
       CDUAtcLatRequestFansA.ShowPage1(mcdu, data);
     };
+    setKeyNotActiveLskActions(mcdu);
   }
 }