+++ fbw-a32nx/src/systems/instruments/src/MCDU/legacy_pages/atsu/FansA/A320_Neo_CDU_ATC_ProcedureRequest.ts
@@ -4,7 +4,7 @@ import { Keypad } from '../../../legacy/A320_Neo_CDU_Keypad';
 import { CDUAtcFlightReq } from '../A320_Neo_CDU_ATC_FlightReq';
 import { CDUAtcTextFansA } from '../FansA/A320_Neo_CDU_ATC_Text';
 import { NXSystemMessages } from '../../../messages/NXSystemMessages';
-import { LegacyAtsuPageInterface } from '../../../legacy/LegacyAtsuPageInterface';
+import { LegacyAtsuPageInterface, setKeyNotActiveLskActions } from '../../../legacy/LegacyAtsuPageInterface';
 
 export class CDUAtcProcedureRequest {
   static CreateDataBlock() {
@@ -53,6 +53,7 @@ export class CDUAtcProcedureRequest {
 
   static ShowPage(mcdu: LegacyAtsuPageInterface, data = CDUAtcProcedureRequest.CreateDataBlock()) {
     mcdu.clearDisplay();
+    mcdu.page.Current = mcdu.page.ATCProcedureRequest;
 
     let sid = '[   ][color]cyan';
     let star = '[   ][color]cyan';
@@ -207,5 +208,6 @@ export class CDUAtcProcedureRequest {
         }
       }
     };
+    setKeyNotActiveLskActions(mcdu);
   }
 }