+++ fbw-a32nx/src/systems/instruments/src/MCDU/legacy_pages/atsu/FansA/A320_Neo_CDU_ATC_VertRequest.ts
@@ -10,7 +10,7 @@ import { Keypad } from '../../../legacy/A320_Neo_CDU_Keypad';
 import { CDUAtcFlightReq } from '../A320_Neo_CDU_ATC_FlightReq';
 import { CDUAtcTextFansA } from '../FansA/A320_Neo_CDU_ATC_Text';
 import { NXSystemMessages } from '../../../messages/NXSystemMessages';
-import { LegacyAtsuPageInterface } from '../../../legacy/LegacyAtsuPageInterface';
+import { LegacyAtsuPageInterface, setKeyNotActiveLskActions } from '../../../legacy/LegacyAtsuPageInterface';
 
 export class CDUAtcVertRequestFansA {
   static CreateDataBlock(): any {
@@ -195,6 +195,7 @@ export class CDUAtcVertRequestFansA {
 
   static ShowPage1(mcdu: LegacyAtsuPageInterface, data = CDUAtcVertRequestFansA.CreateDataBlock()) {
     mcdu.clearDisplay();
+    mcdu.page.currentStation = mcdu.page.ATCVertRequest;
 
     let climbTo = '[   ][color]cyan';
     let descentTo = '[   ][color]cyan';
@@ -447,6 +448,7 @@ export class CDUAtcVertRequestFansA {
 
   static ShowPage2(mcdu: LegacyAtsuPageInterface, data = CDUAtcVertRequestFansA.CreateDataBlock()) {
     mcdu.clearDisplay();
+    mcdu.page.Current = mcdu.page.ATCVertRequest;
 
     let climbStart = '[   ]/[   ][color]cyan';
     if (data.climbStart) {
@@ -580,5 +582,6 @@ export class CDUAtcVertRequestFansA {
     mcdu.onNextPage = () => {
       CDUAtcVertRequestFansA.ShowPage1(mcdu, data);
     };
+    setKeyNotActiveLskActions(mcdu);
   }
 }