+++ fbw-a32nx/src/systems/instruments/src/MCDU/legacy_pages/atsu/FansB/A320_Neo_CDU_ATC_Text.ts
@@ -4,7 +4,7 @@ import { Keypad } from '../../../legacy/A320_Neo_CDU_Keypad';
 import { CDUAtcFlightReq } from '../A320_Neo_CDU_ATC_FlightReq';
 import { CDUAtcMenu } from '../A320_Neo_CDU_ATC_Menu';
 import { NXSystemMessages } from '../../../messages/NXSystemMessages';
-import { LegacyAtsuPageInterface } from '../../../legacy/LegacyAtsuPageInterface';
+import { LegacyAtsuPageInterface, setKeyNotActiveLskActions } from '../../../legacy/LegacyAtsuPageInterface';
 
 export class CDUAtcTextFansB {
   static CreateDataBlock() {
@@ -41,6 +41,7 @@ export class CDUAtcTextFansB {
 
   static ShowPage(mcdu: LegacyAtsuPageInterface, messages, data = CDUAtcTextFansB.CreateDataBlock()) {
     mcdu.clearDisplay();
+    mcdu.page.Current = mcdu.page.ATCText;
 
     let erase = '\xa0ERASE';
     let reqDisplay = 'DCDU\xa0[color]cyan';
@@ -133,5 +134,6 @@ export class CDUAtcTextFansB {
         }
       }
     };
+    setKeyNotActiveLskActions(mcdu);
   }
 }