+++ fbw-a32nx/src/systems/instruments/src/MCDU/legacy_pages/atsu/FansB/A320_Neo_CDU_ATC_UsualRequest.ts
@@ -4,7 +4,7 @@ import { Keypad } from '../../../legacy/A320_Neo_CDU_Keypad';
 import { CDUAtcMenu } from '../A320_Neo_CDU_ATC_Menu';
 import { CDUAtcTextFansB } from '../FansB/A320_Neo_CDU_ATC_Text';
 import { NXSystemMessages } from '../../../messages/NXSystemMessages';
-import { LegacyAtsuPageInterface } from '../../../legacy/LegacyAtsuPageInterface';
+import { LegacyAtsuPageInterface, setKeyNotActiveLskActions } from '../../../legacy/LegacyAtsuPageInterface';
 import { WaypointEntryUtils } from '@fmgc/flightplanning/WaypointEntryUtils';
 
 export class CDUAtcUsualRequestFansB {
@@ -280,5 +280,6 @@ export class CDUAtcUsualRequestFansB {
         }
       }
     };
+    setKeyNotActiveLskActions(mcdu);
   }
 }