+++ fbw-a32nx/src/systems/instruments/src/MCDU/legacy_pages/atsu/FansB/A320_Neo_CDU_ATC_VertRequest.ts
@@ -4,7 +4,7 @@ import { Keypad } from '../../../legacy/A320_Neo_CDU_Keypad';
 import { CDUAtcFlightReq } from '../A320_Neo_CDU_ATC_FlightReq';
 import { CDUAtcTextFansB } from '../FansB/A320_Neo_CDU_ATC_Text';
 import { NXSystemMessages } from '../../../messages/NXSystemMessages';
-import { LegacyAtsuPageInterface } from '../../../legacy/LegacyAtsuPageInterface';
+import { LegacyAtsuPageInterface, setKeyNotActiveLskActions } from '../../../legacy/LegacyAtsuPageInterface';
 
 export class CDUAtcVertRequestFansB {
   static CreateDataBlock(): any {
@@ -49,6 +49,7 @@ export class CDUAtcVertRequestFansB {
 
   static ShowPage(mcdu: LegacyAtsuPageInterface, data = CDUAtcVertRequestFansB.CreateDataBlock()) {
     mcdu.clearDisplay();
+    mcdu.page.Current = mcdu.page.ATCVertRequest;
 
     let climbTo = '[   ][color]cyan';
     let descentTo = '[   ][color]cyan';
@@ -181,5 +182,6 @@ export class CDUAtcVertRequestFansB {
         }
       }
     };
+    setKeyNotActiveLskActions(mcdu);
   }
 }